Adapt data type for python-fcl/fcl 0.6.1
Credit to @aothms #919 
Need testing with python-fc 0.0.12l / fcl 0.5 to make sure it breaks nothing

+++ src/ifcclash/collision.py
@@ -185,7 +185,7 @@ class CollisionManager(object):
         # if no transform passed, assume identity transform
         if transform is None:
             transform = np.eye(4)
-        transform = np.asanyarray(transform, dtype=np.float32)
+        transform = np.asanyarray(transform, dtype=np.float64)
         if transform.shape != (4, 4):
             raise ValueError('transform must be (4,4)!')
